chore(client): move Logo components to assets
