+++ client/src/components/icons/LogoSidedrawer.svelte
