+++ client/src/components/icons/MainLogo.svelte
