+++ client/src/pages/Home.svelte
@@ -9,7 +9,7 @@
     import Camera from '../components/icons/Camera.svelte';
     import Search from '../components/icons/Search.svelte';
     import TextInput from '../components/ui/TextInput.svelte';
-    import MainLogo from '../assets/logo/MainLogo.svelte';
+    import MainLogo from '../components/icons/MainLogo.svelte';
     import Modal from '../components/ui/Modal.svelte';
     import QrScanner from '../components/ui/QRScanner.svelte';
     import { assert } from '../assert.ts';
